Refactor MailRoutes MM-307
