+++ controllers/mail/index.js
@@ -1,0 +1,9 @@
+const { createMailVerification } = require('./createMailVerification');
+const { retryMailVerification } = require('./retryMailVerification');
+const { createCertificateByMailCode } = require('./createCertificateByMailCode');
+
+module.exports = {
+    createMailVerification,
+    retryMailVerification,
+    createCertificateByMailCode,
+};
